Coalesced tests in previous test file structure.
