+++ tests/outlineOTF_test.py
@@ -1,6 +1,7 @@
 from fontTools.ttLib import TTFont
 from defcon import Font
 from ufo2ft.outlineOTF import OutlineTTFCompiler
+from ufo2ft import compileTTF
 import unittest
 import os
 
@@ -43,6 +44,50 @@ class TestOutlineTTCompiler(unittest.TestCase):
         compiler = OutlineTTFCompiler(self.ufo)
         compiler.compile()
         self.assertTrue('gasp' not in compiler.otf)
+
+
+class TestGlyphOrder(unittest.TestCase):
+
+    def setUp(self):
+        self.ufo = getTestUFO()
+
+    def test_compile_original_glyph_order(self):
+        DEFAULT_ORDER = ['.notdef', 'space', 'a', 'b', 'c']
+        compiler = OutlineTTFCompiler(self.ufo)
+        compiler.compile()
+        self.assertEqual(compiler.otf.getGlyphOrder(), DEFAULT_ORDER)
+
+    def test_compile_tweaked_glyph_order(self):
+        NEW_ORDER = ['.notdef', 'space', 'b', 'a', 'c']
+        self.ufo.lib['public.glyphOrder'] = NEW_ORDER
+        compiler = OutlineTTFCompiler(self.ufo)
+        compiler.compile()
+        self.assertEqual(compiler.otf.getGlyphOrder(), NEW_ORDER)
+
+    def test_compile_strange_glyph_order(self):
+        """Move space and .notdef to end of glyph ids
+        ufo2ft always puts .notdef first.
+        """
+        NEW_ORDER = ['b', 'a', 'c', 'space', '.notdef']
+        EXPECTED_ORDER = ['.notdef', 'b', 'a', 'c', 'space']
+        self.ufo.lib['public.glyphOrder'] = NEW_ORDER
+        compiler = OutlineTTFCompiler(self.ufo)
+        compiler.compile()
+        self.assertEqual(compiler.otf.getGlyphOrder(), EXPECTED_ORDER)
+
+
+class TestProductionNames(unittest.TestCase):
+
+    def setUp(self):
+        self.ufo = getTestUFO()
+
+    def test_compile_without_production_names(self):
+        result = compileTTF(self.ufo, useProductionNames=False)
+        self.assertEqual(result.getGlyphOrder(), ['.notdef', 'space', 'a', 'b', 'c'])
+
+    def test_compile_with_production_names(self):
+        result = compileTTF(self.ufo, useProductionNames=True)
+        self.assertEqual(result.getGlyphOrder(), ['.notdef', 'uni0020', 'uni0061', 'uni0062', 'uni0063'])
 
 
 if __name__ == "__main__":
